Separated tests into multiple files
